Add ability to create and read CSV sheet archives (#20)
This is an attempt to create snapshots of all the sheets required so that the pipeline can be run in a more reproducible way.

When sheets are read directly from Google Sheets, there is no knowing whether one or more people may be in the middle of editing one or more of the required sheets and there is a possibility that the one or more sheets may be invalid at the time that the pipeline is run.

The proposed solution is to introduce the ability to download all the Google Sheets required by a pipeline run, save them as CSV files and package them as a zip file i.e. a snapshot. Snapshots should be taken at a time when the Google Sheets are known to be in a good state and are not currently being edited.

A snapshot can created by running: `python -m parenttext_pipeline.archive`

---------

Co-authored-by: Ian Stride <ianstride@gmail.com>