Allow compatible releases of the toolkit
It used to be impossible to install a different version of 'rpft' than the one
specified in this project. This change allows a later version to be installed at
the level of a deployment, for example, which might allow a deployment to
upgrade rpft before it is upgraded in the pipeline. This means that deployments
might be able to use new toolkit features sooner.